backend/hw1/requirements.txt
fastapi 
uvicorn[standart]
sqlmodel
python-multipart
passlib[bcrypt]
python-jose[cryptography] 
psycopg2-binary
python-dotenv